contrib/backend/requirements.txt
flask
together
pillow